Fixed some FindBugs issues, removed dead code etc...

+++ gameengine-lua/src/main/java/org/luaj/vm2/Lua.java
@@ -147,7 +147,7 @@ public class Lua {
 
 	/** gets the index of the constant */
 	public static int INDEXK(int r) {
-		return ((int)(r) & ~BITRK);
+		return (r & ~BITRK);
 	}
 
 	public static final int MAXINDEXRK	= (BITRK - 1);
